refactor(portal): import shared atoms from scenes/common directly, drop re-export forwarding
Every consumer of createAppDialogOpenedAtom and banMessageDialogOpenedAtom now
imports the atom from its scenes/common module explicitly, instead of going
through a re-export in the file the atom was split out of. Same single atom
instance either way — this just makes every import path point at the real
home. The BanMessageDialog fork carries the identical edit, so the copies stay
byte-identical.
