+++ web/scenes/Portal/layout/AppSelector/index.tsx
@@ -26,7 +26,7 @@ import clsx from "clsx";
 import { useAtom } from "jotai";
 import { useParams, useRouter } from "next/navigation";
 import { useCallback, useMemo } from "react";
-import { createAppDialogOpenedAtom } from "../Header";
+import { createAppDialogOpenedAtom } from "@/scenes/common/layout/Header/atoms";
 
 export const AppSelector = () => {
   const router = useRouter();
